Fixed a bug with cycles
Previously, I was doing path position lookups by segment ID, but because
this is not an acyclic graph, segment IDs can be present in the same
path multiple times, resulting in the same segment having multiple
positions in the path. So, I switched to using the path index to do
lookups, which required keeping track of that, which took a bit of
rewriting.